refactor(plugin): Consolidate duplicated URL helpers into targets package
Move parseRepoURL, sanitizeRepoName, sanitizeForFilename, and
repoDisplayName into the targets package as exported functions
(ParseRepoURL, SanitizeRepoURL, RepoDisplayName) to eliminate
duplication across scan and results packages.
